+++ src/app/components/main-window/sidebar/logout-btn/logout-btn.component.html
